+++ src/errors.rs
@@ -15,8 +15,14 @@ pub enum Error {
     #[fail(display = "failed to bind to '{}'", _0)]
     BindAddress(String),
 
+    #[fail(display = "could not get jail name")]
+    JailError(#[fail(cause)] jail::JailError),
+
     #[fail(display = "jail_exporter must be run as root")]
     NotRunningAsRoot,
+
+    #[fail(display = "error within Prometheus library")]
+    PrometheusError(#[fail(cause)] prometheus::Error),
 
     #[fail(display = "RACCT/RCTL: {}", _0)]
     RctlUnavailable(String),
@@ -26,6 +32,7 @@ pub enum Error {
 
     #[fail(display = "Could not parse SocketAddr: {}", _0)]
     SocketAddr(String),
+
 }
 
 // Implements basic output, allowing the above display strings to be used when
@@ -33,5 +40,17 @@ pub enum Error {
 impl fmt::Debug for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self)
+    }
+}
+
+impl From<jail::JailError> for Error {
+    fn from(e: jail::JailError) -> Self {
+        Error::JailError(e)
+    }
+}
+
+impl From<prometheus::Error> for Error {
+    fn from(e: prometheus::Error) -> Self {
+        Error::PrometheusError(e)
     }
 }